BJS2_31377 : Changed the architecture app Delete class BJS2_31377 and create package bjs2_31377.
