+++ src/main/java/school/faang/Main.java
@@ -2,6 +2,6 @@ package school.faang;
 
 public class Main {
 	public static void main(String[] args) {
-		new Tasks.BJS2_31377().run(); //
+
 	}
 }
